+++ src/cli/deploy.ts
@@ -2,6 +2,7 @@ import PackageDeployerConfiguration from "@/packageDeployer/PackageDeployerConfi
 import { PackageDeployer } from "..";
 import NodePackageList from "@/package/NodePackageList";
 import RemotePackageList from "@/package/RemotePackageList";
+import PackageDeployerOrchestrator from "@/packageDeployer/PackageDeployerOrchestrator";
 
 /**
  * Deploy all packages
@@ -47,8 +48,8 @@ export default async function deployMain(
 			if (registryPassword && registryUsername && incrementalEnabled) {
 				console.log(`Smart(Incremental) package deployment`);
 
-				// Package deployer
-				const pkgDeployer = new PackageDeployer(
+				// Orchestrator
+				const orchestrator = new PackageDeployerOrchestrator(
 					config,
 					packageList,
 					remotePackageList,
@@ -56,15 +57,15 @@ export default async function deployMain(
 						ignoreApps,
 					}
 				);
-				await pkgDeployer.incrementalDeployment();
+				await orchestrator.incrementalDeployment();
 			} else {
 				console.log(
 					`No registry password nor username, defaulting to deploying all at once.\n`,
 					`Make sure you set the registry username and password so that updates\n`,
 					`are incremental, and you don't re-build what you already had.`
 				);
-				// Initialize package deployer and deploy all
-				const pkgDeployer = new PackageDeployer(
+				// Orchestrator
+				const orchestrator = new PackageDeployerOrchestrator(
 					config,
 					packageList,
 					remotePackageList,
@@ -72,7 +73,7 @@ export default async function deployMain(
 						ignoreApps,
 					}
 				);
-				await pkgDeployer.deploy();
+				await orchestrator.deploy();
 			}
 		}
 	);