+++ src/packageDeployer/PackageDeployer.ts
@@ -1,171 +1,32 @@
 import path from "path";
 import fsp from "fs/promises";
-import semver from "semver";
-import pc from "picocolors";
 
-import PackageDeployerConfiguration from "./PackageDeployerConfiguration";
 import DefaultConfigFolder from "../configuration/DefaultConfigFolder";
-import { IRemotePackageInfo, ITaskDeploymentResult } from "../types";
-import KhansDependencyGraph from "@/graph/KhansDependencyGraph";
+import { ITaskDeploymentResult } from "../types";
 import NodePackage from "@/package/NodePackage";
-import NodePackageList from "@/package/NodePackageList";
-import RemotePackageList from "@/package/RemotePackageList";
 
 /**
  * Package deployer
  */
 export default class PackageDeployer {
-	config: PackageDeployerConfiguration;
-	remotePackageList: RemotePackageList;
-	nodePackagesList: NodePackageList;
-	whitelist: Array<string> = [];
-
-	// When building ignore apps
-	// Nice feature when you just need to update packages
-	ignoreApps: boolean = false;
+	packages: Array<NodePackage> = [];
 
 	/**
 	 * Package deployer constructor
 	 *
 	 * @param config
 	 */
-	constructor(
-		config: PackageDeployerConfiguration,
-		nodePackages: NodePackageList,
-		remotePackageList: RemotePackageList,
-		options?: {
-			whitelist?: Array<string>;
-			ignoreApps?: boolean;
-		}
-	) {
-		this.config = config;
-		this.nodePackagesList = nodePackages;
-		this.remotePackageList = remotePackageList;
-
-		// Set options
-		if (options) {
-			// Whitelist
-			if (options.whitelist) {
-				this.whitelist = options.whitelist;
-			}
-
-			// Ignore apps
-			if (options.ignoreApps) {
-				this.ignoreApps = options.ignoreApps;
-			}
-		}
-	}
-
-	/**
-	 * Get node packages
-	 *
-	 * Call this only after fetching node packages
-	 */
-	getNodePackages() {
-		// Whitelisted packages
-		const whitelistedPackages =
-			this.config.configuration.repositoriesListing.use === "whitelist"
-				? this.nodePackagesList.getNodePackagesFolderNameWhitelisted(
-						this.config.getWhitelist()
-				  )
-				: this.nodePackagesList.getNodePackagesFolderNameBlacklisted(
-						this.config.getBlacklist()
-				  );
-
-		// Filter out applications if the argument was given
-		const packages = whitelistedPackages.filter((pkg) => {
-			// Check if the user flagged ignore apps as true
-			if (this.ignoreApps) {
-				// Check if the package is private
-				return !pkg.packageJson.private;
-			}
-
-			// Otherwise all of them can pass
-			return true;
-		});
-
-		return packages;
-	}
-
-	/**
-	 * Get remote packages
-	 */
-	getRemotePackages() {
-		return this.remotePackageList.getPackages();
-	}
-
-	/**
-	 * Get build order
-	 *
-	 * Make sure that node packages exist
-	 */
-	getBuildOrder() {
-		// Get the dependency graph
-		const dependencyGraph = new KhansDependencyGraph(
-			this.getNodePackages()
-		);
-		const buildOrder = dependencyGraph.getBuildOrder();
-
-		// Filter out those that aren't in the whitelist
-		const finalBuildOrder =
-			this.whitelist.length > 0
-				? buildOrder.filter((pkg) =>
-						this.whitelist.includes(pkg.packageName)
-				  )
-				: buildOrder;
-
-		return finalBuildOrder;
-	}
-
-	/**
-	 * TODO: Has to be removed from this class
-	 * Get incremental build order
-	 *
-	 * Make sure both, node packages and remote packages were fetch.
-	 */
-	getIncrementalBuildOrder() {
-		const nodePackages = this.getNodePackages();
-		const remotePackages = this.getRemotePackages();
-
-		// The new whitelist will be the new packages
-		const directlyAffectedNames = nodePackages
-			.filter((pkg) => {
-				// Get the remote package info
-				const remotePkgInfo = remotePackages.get(pkg.packageName);
-
-				// The package is not deployed
-				if (!remotePkgInfo) {
-					return true;
-				}
-
-				// Simple check, check if both versions differ
-				return semver.gt(pkg.version, remotePkgInfo.version);
-			})
-			// Just get the package names
-			.map((pkg) => pkg.packageName);
-
-		// Use the Graph to find "Transitive" dependents
-		// Even if App-B didn't change version, if Core-A (its dependency) changed,
-		// App-B needs a redeploy.
-		const graph = new KhansDependencyGraph(nodePackages);
-		const finalBuildOrder = graph.getAffectedPackages(
-			directlyAffectedNames
-		);
-
-		return finalBuildOrder;
+	constructor(nodePackages: Array<NodePackage>) {
+		this.packages = nodePackages;
 	}
 
 	/**
 	 * Deploy
 	 */
-	async deploy(customBuildOrder?: Array<NodePackage>) {
-		const buildOrder = customBuildOrder
-			? customBuildOrder
-			: this.getBuildOrder();
-
+	async deploy() {
 		// Deploy all packages
 		const packageDeploymentResult: Array<ITaskDeploymentResult> = [];
-		for (const nodePackage of buildOrder) {
+		for (const nodePackage of this.packages) {
 			try {
 				await nodePackage.install();
 				await nodePackage.build();
@@ -179,6 +40,7 @@ export default class PackageDeployer {
 				packageDeploymentResult.push({
 					packageName: nodePackage.packageName,
 					name: nodePackage.name,
+					version: nodePackage.version,
 					success: true,
 				});
 			} catch (err) {
@@ -188,6 +50,7 @@ export default class PackageDeployer {
 				packageDeploymentResult.push({
 					packageName: nodePackage.packageName,
 					name: nodePackage.name,
+					version: nodePackage.version,
 					success: false,
 				});
 			}
@@ -195,32 +58,8 @@ export default class PackageDeployer {
 
 		// Save as json
 		this.saveDeploymentResult(packageDeploymentResult);
-	}
 
-	/**
-	 * TODO: Has to be removed from this class
-	 * Incremental deployment
-	 *
-	 * This will ignore the given whitelist it will make its own
-	 */
-	async incrementalDeployment() {
-		// Get incremental build order and if it's zero return
-		const incrementalBuildOrder = this.getIncrementalBuildOrder();
-		if (incrementalBuildOrder.length === 0) {
-			console.log(
-				pc.green("✅ All packages are up to date. Nothing to deploy.")
-			);
-			return;
-		}
-
-		// Final build order and whitelist
-		const whitelist = incrementalBuildOrder.map((pkg) => pkg.packageName);
-		console.log(
-			`🚀 Packages to deploy in order: ${whitelist.join(" => ")}`
-		);
-
-		// Initialize package deployer and deploy all
-		await this.deploy(incrementalBuildOrder);
+		return packageDeploymentResult;
 	}
 
 	/**